i do not know

+++ TheTop.Application/Entities/Contract.cs
@@ -6,7 +6,7 @@ namespace TheTop.Application.Entities
     {
         public int ContractId { get; set;  }
         public decimal HourSalary { get; set; }
-        public float MonthlyWorkingHours { get; set; }
+        public int MonthlyWorkingHours { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
         public DateTime UpdatedAt { get; set; }
     }
